Version: 0.1.63 Branch: master Date: 25.06.2017 Author: AlexMerlyan
Working: refactoring service layer, insert,update and delete methods working
Changes: refactoring service layer, insert,update and delete methods working
Next: refactoring get methods on service layer

Comment: fuck yeah!

+++ src/main/java/projectpackage/service/maintenanceservice/JournalRecordServiceImpl.java
@@ -5,19 +5,21 @@ import lombok.extern.log4j.Log4j;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import projectpackage.dto.IUDAnswer;
 import projectpackage.model.maintenances.Complimentary;
 import projectpackage.model.maintenances.JournalRecord;
-import projectpackage.dto.IUDAnswer;
 import projectpackage.model.maintenances.Maintenance;
 import projectpackage.model.orders.Order;
+import projectpackage.repository.maintenancedao.JournalRecordDAO;
 import projectpackage.repository.ordersdao.OrderDAO;
 import projectpackage.repository.support.daoexceptions.DeletedObjectNotExistsException;
 import projectpackage.repository.support.daoexceptions.ReferenceBreakException;
-import projectpackage.repository.support.daoexceptions.TransactionException;
 import projectpackage.repository.support.daoexceptions.WrongEntityIdException;
-import projectpackage.repository.maintenancedao.JournalRecordDAO;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 /**
  * Created by Dima on 21.05.2017.
@@ -58,31 +60,33 @@ public class JournalRecordServiceImpl implements JournalRecordService{
     }
 
     @Override
-    public JournalRecord getSingleEntityById(int id) {
+    public JournalRecord getSingleEntityById(Integer id) {
         JournalRecord journalRecord = journalRecordDAO.getJournalRecord(id);
         if (null == journalRecord) LOGGER.info("Returned NULL!!!");
         return journalRecord;
     }
 
     @Override
-    public IUDAnswer deleteJournalRecord(int id) {
+    public IUDAnswer deleteJournalRecord(Integer id) {
+        if (id == null) return new IUDAnswer(false, NULL_ID);
         try {
             journalRecordDAO.deleteJournalRecord(id);
         } catch (ReferenceBreakException e) {
-            LOGGER.warn("Entity has references on self", e);
-            return new IUDAnswer(id,false, e.printReferencesEntities());
+            return journalRecordDAO.rollback(id, ON_ENTITY_REFERENCE, e);
         } catch (DeletedObjectNotExistsException e) {
-            LOGGER.warn("Entity with that id does not exist!", e);
-            return new IUDAnswer(id, "deletedObjectNotExists");
+            return journalRecordDAO.rollback(id, DELETED_OBJECT_NOT_EXISTS, e);
         } catch (WrongEntityIdException e) {
-            LOGGER.warn("This id belong another entity class!", e);
-            return new IUDAnswer(id, "wrongDeleteId");
+            return journalRecordDAO.rollback(id, WRONG_DELETED_ID, e);
+        } catch (IllegalArgumentException e) {
+            return journalRecordDAO.rollback(id, NULL_ID, e);
         }
+        journalRecordDAO.commit();
         return new IUDAnswer(id, true);
     }
 
     @Override
     public IUDAnswer insertJournalRecord(JournalRecord journalRecord) {
+        if (journalRecord == null) return null;
         Order order = orderDAO.getOrder(journalRecord.getOrderId());
         Set<Complimentary> freeComplimentaries = order.getCategory().getComplimentaries();
         Set<Maintenance> freeMaintenances = new HashSet<>();
@@ -97,24 +101,25 @@ public class JournalRecordServiceImpl implements JournalRecordService{
         Integer journalRecordId = null;
         try {
             journalRecordId = journalRecordDAO.insertJournalRecord(journalRecord);
-            LOGGER.info("Get from DB journalRecordId = " + journalRecordId);
-        } catch (TransactionException e) {
-            LOGGER.warn("Catched transactionException!!!", e);
-            return new IUDAnswer(journalRecordId,false, "transactionInterrupt");
+        } catch (IllegalArgumentException e) {
+            return journalRecordDAO.rollback(WRONG_FIELD, e);
         }
+        journalRecordDAO.commit();
         return new IUDAnswer(journalRecordId,true);
     }
 
     @Override
-    public IUDAnswer updateJournalRecord(int id, JournalRecord newJournalRecord) {
+    public IUDAnswer updateJournalRecord(Integer id, JournalRecord newJournalRecord) {
+        if (newJournalRecord == null) return null;
+        if (id == null) return new IUDAnswer(false, NULL_ID);
         try {
             newJournalRecord.setObjectId(id);
             JournalRecord oldJournalRecord = journalRecordDAO.getJournalRecord(id);
             journalRecordDAO.updateJournalRecord(newJournalRecord, oldJournalRecord);
-        } catch (TransactionException e) {
-            LOGGER.warn("Catched transactionException!!!", e);
-            return new IUDAnswer(id,false, "transactionInterrupt");
+        } catch (IllegalArgumentException e) {
+            return journalRecordDAO.rollback(WRONG_FIELD, e);
         }
+        journalRecordDAO.commit();
         return new IUDAnswer(id,true);
     }
 }
