+++ src/main/java/projectpackage/service/roomservice/RoomTypeServiceImpl.java
@@ -10,12 +10,11 @@ import projectpackage.model.rooms.RoomType;
 import projectpackage.repository.roomsdao.RoomTypeDAO;
 import projectpackage.repository.support.daoexceptions.DeletedObjectNotExistsException;
 import projectpackage.repository.support.daoexceptions.ReferenceBreakException;
-import projectpackage.repository.support.daoexceptions.TransactionException;
 import projectpackage.repository.support.daoexceptions.WrongEntityIdException;
 import projectpackage.service.orderservice.CategoryService;
 
-import java.util.Date;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -67,7 +66,6 @@ public class RoomTypeServiceImpl implements RoomTypeService{
             long days = (finishDate.getTime() - startDate.getTime()) / (24 * 60 * 60 * 1000);
             Long categoryCost = categoryPrice * days;
             orderDTO.setCategoryCost(categoryCost);
-            //LOGGER.info( + "****************************************************");
             orderDTO.setLivingPersons(numberOfPeople);
             orderDTO.setRoomTypeId(roomType.getObjectId());
             orderDTO.setArrival(startDate);
@@ -98,52 +96,55 @@ public class RoomTypeServiceImpl implements RoomTypeService{
     }
 
     @Override
-    public RoomType getSingleRoomTypeById(int id) {
+    public RoomType getSingleRoomTypeById(Integer id) {
         RoomType roomType = roomTypeDAO.getRoomType(id);
         if (roomType == null) LOGGER.info("Returned NULL!!!");
         return roomType;
     }
 
     @Override
-    public IUDAnswer deleteRoomType(int id) {
+    public IUDAnswer deleteRoomType(Integer id) {
+        if (id == null) return new IUDAnswer(false, NULL_ID);
         try {
             roomTypeDAO.deleteRoomType(id);
         } catch (ReferenceBreakException e) {
-            LOGGER.warn("Entity has references on self", e);
-            return new IUDAnswer(id,false, e.printReferencesEntities());
+            return roomTypeDAO.rollback(id, ON_ENTITY_REFERENCE, e);
         } catch (DeletedObjectNotExistsException e) {
-            LOGGER.warn("Entity with that id does not exist!", e);
-            return new IUDAnswer(id, "deletedObjectNotExists");
+            return roomTypeDAO.rollback(id, DELETED_OBJECT_NOT_EXISTS, e);
         } catch (WrongEntityIdException e) {
-            LOGGER.warn("This id belong another entity class!", e);
-            return new IUDAnswer(id, "wrongDeleteId");
+            return roomTypeDAO.rollback(id, WRONG_DELETED_ID, e);
+        } catch (IllegalArgumentException e) {
+            return roomTypeDAO.rollback(id, NULL_ID, e);
         }
+        roomTypeDAO.commit();
         return new IUDAnswer(id, true);
     }
 
     @Override
     public IUDAnswer insertRoomType(RoomType roomType) {
+        if (roomType == null) return null;
         Integer roomTypeId = null;
         try {
             roomTypeId = roomTypeDAO.insertRoomType(roomType);
-            LOGGER.info("Get from DB roomId = " + roomTypeId);
-        } catch (TransactionException e) {
-            LOGGER.warn("Catched transactionException!!!", e);
-            return new IUDAnswer(roomTypeId,false, "transactionInterrupt");
+        } catch (IllegalArgumentException e) {
+            return roomTypeDAO.rollback(WRONG_FIELD, e);
         }
+        roomTypeDAO.commit();
         return new IUDAnswer(roomTypeId,true);
     }
 
     @Override
-    public IUDAnswer updateRoomType(int id, RoomType newRoomType) {
+    public IUDAnswer updateRoomType(Integer id, RoomType newRoomType) {
+        if (newRoomType == null) return null;
+        if (id == null) return new IUDAnswer(false, NULL_ID);
         try {
             newRoomType.setObjectId(id);
             RoomType oldRoomType = roomTypeDAO.getRoomType(id);
             roomTypeDAO.updateRoomType(newRoomType, oldRoomType);
-        } catch (TransactionException e) {
-            LOGGER.warn("Catched transactionException!!!", e);
-            return new IUDAnswer(id,false, "transactionInterrupt");
+        } catch (IllegalArgumentException e) {
+            return roomTypeDAO.rollback(WRONG_FIELD, e);
         }
+        roomTypeDAO.commit();
         return new IUDAnswer(id,true);
     }
 }
